Move menus into menus folder
