+++ trview.app/Menus/LevelSwitcher.h
