+++ trview.app/Menus/RecentFiles.cpp
