+++ trview.app/Menus/RecentFiles.h
