+++ trview.app/Menus/ViewMenu.cpp
@@ -1,5 +1,5 @@
 #include "ViewMenu.h"
-#include "WindowIDs.h"
+#include <trview.app/WindowIDs.h>
 
 namespace trview
 {
