+++ trview.app/Menus/ViewMenu.h
